+++ master/src/main/scala/org/apache/celeborn/service/deploy/master/http/api/v1/ApiV1MasterResource.scala
@@ -37,4 +37,7 @@ class ApiV1MasterResource extends ApiRequestContext {
 
   @Path("ratis")
   def ratis: Class[RatisResource] = classOf[RatisResource]
+
+  @Path("ratis/local")
+  def ratisLocal: Class[RatisLocalResource] = classOf[RatisLocalResource]
 }
